refactor: enhance reusable component order list
